Fixed issue with overflow in getNumberBufAsLong
getNumberBufAsLong was using a long to extract the bits from the buffer to get the necessary value. However, problems will happen when you are trying to extract a 64-bit field that spans 9 bytes. One of the bytes will be lost when trying to shove it all in a long. Therefore, I rewrote the method to return a BigInteger and did all the math using the BigInteger so the ninth byte will no longer be lost.